Coordenada na ActionBar

+++ src/main/java/com/henriquenapimo1/ping/SuperPing.java
@@ -147,7 +147,7 @@ public final class SuperPing extends JavaPlugin implements Listener {
                     p.sendActionBar(Component.empty().append(
                             Component.text("§bAlvo: ")
                                     .append(Component.translatable(e.getType().translationKey()).color(TextColor.color(NamedTextColor.AQUA))
-                                            .append(Component.text(" | Distância: "+((int) distText)+" m (X: "+e.getX()+", Y: "+e.getY()+", Z: "+e.getZ()+")")))));
+                                            .append(Component.text(" | Distância: "+((int) distText)+" m (X: "+((int) e.getX())+", Y: "+((int) e.getY())+", Z: "+((int) e.getZ())+")")))));
 
                     counter += 1;
                     if(counter >= 100) {
